Update Proxy.py
Create a server socket and bind the server socket to a host and port. Set the maximum number of connections (for example, 6) and start listening. In the while True loop, the client connection is accepted and a new socket object is fetched. Receives data from a client socket, using the recv method. These steps are added comments as well.