Update Proxy.py
Create a socket to connect to origin server，and store in originServerSocket. Get the IP address for a hostname. Create origin server request line and headers to send and store in originServerRequestHeader and originServerRequest originServerRequest is the first line in the request and originServerRequestHeader is the second line in the request. Add a While Loop from line 181. Add codes for sending the response to the client (line 190). At the end, add codes for saving origin server response in the cache file.